test: add custom module and fn isolation fixtures
module_isolation fixture completely wipes out the chain resetting to
block 0, essentially borking all session level fixtures that are
created. Alternative was to store the height of the chain at the
beginning of the module and then reset to that point at the end of the
module ... will see how this turns out.

Realistically session level fixtures should just be deployments that are
uninitialized. Any modifying fixtures should be at the module level.